lerobot/notes_siyu/vlm_fine_tune_example/requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
jupyter>=1.0.0
qwen-vl-utils
accelerate>=0.20.0
